split out steplist thunks and clean up
