+++ protocol-designer/src/steplist/actions/index.js
@@ -3,6 +3,7 @@ import handleFormChange from './handleFormChange'
 
 export * from './actions'
 export * from './types'
+export * from './thunks'
 
 export {
   handleFormChange,
