+++ protocol-designer/src/steplist/selectors.js
@@ -9,7 +9,6 @@ import {selectors as labwareIngredSelectors} from '../labware-ingred/reducers'
 import {
   getFormWarnings,
   getFormErrors,
-  generateNewForm,
   stepFormToArgs,
 } from './formLevel'
 import type {FormError, FormWarning} from './formLevel'
@@ -27,7 +26,6 @@ import type {
 
 import type {
   FormData,
-  BlankForm,
   StepIdType,
 } from '../form-types'
 
@@ -211,35 +209,6 @@ const stepCreationButtonExpandedSelector: Selector<boolean> = createSelector(
   (state: RootState) => state.stepCreationButtonExpanded
 )
 
-type MakeGetSelectedStepFormData = ({defaultNextPipette: string}) => Selector<boolean | FormData | BlankForm>
-const makeGetSelectedStepFormData: MakeGetSelectedStepFormData = ({defaultNextPipette}) => createSelector(
-  getSavedForms,
-  getSelectedStepId,
-  getSteps,
-  (savedStepForms, selectedStepId, steps) => {
-    if (selectedStepId == null) {
-      // no step selected
-      return false
-    }
-
-    if (!steps[selectedStepId]) {
-      console.error(`Step id ${selectedStepId} not in 'steps', could not get form data`)
-      return false
-    }
-
-    return (
-      // existing form
-      savedStepForms[selectedStepId] ||
-      // new blank form
-      generateNewForm({
-        stepId: selectedStepId,
-        stepType: steps[selectedStepId].stepType,
-        defaultNextPipette,
-      })
-    )
-  }
-)
-
 const nextStepId: Selector<number> = createSelector( // generates the next step ID to use
   getSteps,
   (_steps): number => {
@@ -349,7 +318,6 @@ export default {
   getHoveredStepId,
   getActiveItem,
   getHoveredSubstep,
-  makeGetSelectedStepFormData,
   getUnsavedForm,
   formData, // TODO: remove after sunset
   formModalData,
